// src/lib/saleor/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /**
   * The `Date` scalar type represents a Date
   * value as specified by
   * [iso8601](https://en.wikipedia.org/wiki/ISO_8601).
   */
  Date: string;
  /**
   * The `DateTime` scalar type represents a DateTime
   * value as specified by
   * [iso8601](https://en.wikipedia.org/wiki/ISO_8601).
   */
  DateTime: string;
  /**
   * The `GenericScalar` scalar type represents a generic
   * GraphQL scalar value that could be:
   * String, Boolean, Int, Float, List or Object.
   */
  GenericScalar: unknown;
  JSONString: string;
  /**
   * Metadata is a map of key-value pairs, both keys and values are `String`.
   *
   * Example:
   * ```
   * {
   *     "key1": "value1",
   *     "key2": "value2"
   * }
   * ```
   */
  Metadata: Record<string, string>;
  /**
   * Nonnegative Decimal scalar implementation.
   *
   * Should be used in places where value must be nonnegative (0 or greater).
   */
  PositiveDecimal: number;
  UUID: string;
  /** Variables of this type must be set to null in mutations. They will be replaced with a filename from a following multipart part containing a binary file. See: https://github.com/jaydenseric/graphql-multipart-request-spec. */
  Upload: unknown;
  WeightScalar: number;
  /** _Any value scalar as defined by Federation spec. */
  _Any: unknown;
};
// ...
export type MetadataItem = {
  /** Key of a metadata item. */
  key: Scalars['String'];
  /** Value of a metadata item. */
  value: Scalars['String'];
};
// ...
export type Node = {
  /** The ID of the object. */
  id: Scalars['ID'];
};

export type ObjectWithMetadata = {
  /** List of public metadata items. Can be accessed without permissions. */
  metadata: Array<MetadataItem>;
  /**
   * A single key from public metadata.
   *
   * Tip: Use GraphQL aliases to fetch multiple keys.
   */
  metafield?: Maybe<Scalars['String']>;
  /** Public metadata. Use `keys` to control which fields you want to include. The default is to include everything. */
  metafields?: Maybe<Scalars['Metadata']>;
  /** List of private metadata items. Requires staff permissions to access. */
  privateMetadata: Array<MetadataItem>;
  /**
   * A single key from private metadata. Requires staff permissions to access.
   *
   * Tip: Use GraphQL aliases to fetch multiple keys.
   */
  privateMetafield?: Maybe<Scalars['String']>;
  /** Private metadata. Requires staff permissions to access. Use `keys` to control which fields you want to include. The default is to include everything. */
  privateMetafields?: Maybe<Scalars['Metadata']>;
};
// ...
export enum ProductMediaType {
  Image = 'IMAGE',
  Video = 'VIDEO'
}
// ...
export type ProductVariant = Node &
  ObjectWithMetadata & {
    /** List of attributes assigned to this variant. */
    attributes: Array<SelectedAttribute>;
    /** Channel given to retrieve this product variant. Also used by federation gateway to resolve this object in a federated query. */
    channel?: Maybe<Scalars['String']>;
    /**
     * List of price information in channels for the product.
     *
     * Requires one of the following permissions: AUTHENTICATED_APP, AUTHENTICATED_STAFF_USER.
     */
    channelListings?: Maybe<Array<ProductVariantChannelListing>>;
    created: Scalars['DateTime'];
    /**
     * Digital content for the product variant.
     *
     * Requires one of the following permissions: MANAGE_PRODUCTS.
     */
    digitalContent?: Maybe<DigitalContent>;
    /**
     * External ID of this product.
     *
     * Added in Saleor 3.10.
     */
    externalReference?: Maybe<Scalars['String']>;
    id: Scalars['ID'];
    /**
     * List of images for the product variant.
     * @deprecated This field will be removed in Saleor 4.0. Use the `media` field instead.
     */
    images?: Maybe<Array<ProductImage>>;
    /** Gross margin percentage value. */
    margin?: Maybe<Scalars['Int']>;
    /** List of media for the product variant. */
    media?: Maybe<Array<ProductMedia>>;
    /** List of public metadata items. Can be accessed without permissions. */
    metadata: Array<MetadataItem>;
    /**
     * A single key from public metadata.
     *
     * Tip: Use GraphQL aliases to fetch multiple keys.
     *
     * Added in Saleor 3.3.
     *
     * Note: this API is currently in Feature Preview and can be subject to changes at later point.
     */
    metafield?: Maybe<Scalars['String']>;
    /**
     * Public metadata. Use `keys` to control which fields you want to include. The default is to include everything.
     *
     * Added in Saleor 3.3.
     *
     * Note: this API is currently in Feature Preview and can be subject to changes at later point.
     */
    metafields?: Maybe<Scalars['Metadata']>;
    name: Scalars['String'];
    /**
     * Preorder data for product variant.
     *
     * Added in Saleor 3.1.
     *
     * Note: this API is currently in Feature Preview and can be subject to changes at later point.
     */
    preorder?: Maybe<PreorderData>;
    /** Lists the storefront variant's pricing, the current price and discounts, only meant for displaying. */
    pricing?: Maybe<VariantPricingInfo>;
    /** List of private metadata items. Requires staff permissions to access. */
    privateMetadata: Array<MetadataItem>;
    /**
     * A single key from private metadata. Requires staff permissions to access.
     *
     * Tip: Use GraphQL aliases to fetch multiple keys.
     *
     * Added in Saleor 3.3.
     *
     * Note: this API is currently in Feature Preview and can be subject to changes at later point.
     */
    privateMetafield?: Maybe<Scalars['String']>;
    /**
     * Private metadata. Requires staff permissions to access. Use `keys` to control which fields you want to include. The default is to include everything.
     *
     * Added in Saleor 3.3.
     *
     * Note: this API is currently in Feature Preview and can be subject to changes at later point.
     */
    privateMetafields?: Maybe<Scalars['Metadata']>;
    product: Product;
    /** Quantity of a product available for sale in one checkout. Field value will be `null` when no `limitQuantityPerCheckout` in global settings has been set, and `productVariant` stocks are not tracked. */
    quantityAvailable?: Maybe<Scalars['Int']>;
    quantityLimitPerCustomer?: Maybe<Scalars['Int']>;
    /**
     * Total quantity ordered.
     *
     * Requires one of the following permissions: MANAGE_PRODUCTS.
     */
    quantityOrdered?: Maybe<Scalars['Int']>;
    /**
     * Total revenue generated by a variant in given period of time. Note: this field should be queried using `reportProductSales` query as it uses optimizations suitable for such calculations.
     *
     * Requires one of the following permissions: MANAGE_PRODUCTS.
     */
    revenue?: Maybe<TaxedMoney>;
    sku?: Maybe<Scalars['String']>;
    /**
     * Stocks for the product variant.
     *
     * Requires one of the following permissions: MANAGE_PRODUCTS, MANAGE_ORDERS.
     */
    stocks?: Maybe<Array<Stock>>;
    trackInventory: Scalars['Boolean'];
    /** Returns translated product variant fields for the given language code. */
    translation?: Maybe<ProductVariantTranslation>;
    updatedAt: Scalars['DateTime'];
    weight?: Maybe<Weight>;
  };
// ...
export type ProductVariantTranslation = Node & {
  id: Scalars['ID'];
  /** Translation language. */
  language: LanguageDisplay;
  name: Scalars['String'];
};
// ...
export type SelectedAttribute = {
  /** Name of an attribute displayed in the interface. */
  attribute: Attribute;
  /** Values of an attribute. */
  values: Array<AttributeValue>;
};
// ...
export type CheckoutDeleteLineMutationVariables = Exact<{
  checkoutId: Scalars['ID'];
  lineIds: Array<Scalars['ID']> | Scalars['ID'];
}>;

export type CheckoutDeleteLineMutation = {
  checkoutLinesDelete?: {
    errors: Array<{ code: CheckoutErrorCode; message?: string | null; field?: string | null }>;
    checkout?: {
      id: string;
      quantity: number;
      totalPrice: {
        gross: { currency: string; amount: number };
        tax: { currency: string; amount: number };
      };
      subtotalPrice: { gross: { currency: string; amount: number } };
      lines: Array<{
        id: string;
        quantity: number;
        variant: {
          id: string;
          name: string;
          product: {
            id: string;
            slug: string;
            name: string;
            isAvailableForPurchase?: boolean | null;
            description?: string | null;
            seoTitle?: string | null;
            seoDescription?: string | null;
            updatedAt: string;
            pricing?: {
              priceRange?: {
                start?: { gross: { currency: string; amount: number } } | null;
                stop?: { gross: { currency: string; amount: number } } | null;
              } | null;
            } | null;
            media?: Array<{ url: string; type: ProductMediaType; alt: string }> | null;
            collections?: Array<{ name: string }> | null;
            variants?: Array<{
              id: string;
              name: string;
              attributes: Array<{
                attribute: {
                  slug?: string | null;
                  name?: string | null;
                  choices?: { edges: Array<{ node: { name?: string | null } }> } | null;
                };
                values: Array<{ name?: string | null }>;
              }>;
              pricing?: { price?: { gross: { currency: string; amount: number } } | null } | null;
            }> | null;
          };
          attributes: Array<{
            attribute: {
              slug?: string | null;
              name?: string | null;
              choices?: { edges: Array<{ node: { name?: string | null } }> } | null;
            };
            values: Array<{ name?: string | null }>;
          }>;
          pricing?: { price?: { gross: { currency: string; amount: number } } | null } | null;
        };
      }>;
    } | null;
  } | null;
};
// ...
export class TypedDocumentString<TResult, TVariables>
  extends String
  implements DocumentTypeDecoration<TResult, TVariables>
{
  __apiType?: DocumentTypeDecoration<TResult, TVariables>['__apiType'];

  constructor(private value: string, public __meta__?: { hash: string }) {
    super(value);
  }

  toString(): string & DocumentTypeDecoration<TResult, TVariables> {
    return this.value;
  }
}
// ...
export const CheckoutDeleteLineDocument = new TypedDocumentString(`
    mutation CheckoutDeleteLine($checkoutId: ID!, $lineIds: [ID!]!) {
  checkoutLinesDelete(id: $checkoutId, linesIds: $lineIds) {
    errors {
      code
      message
      field
    }
    checkout {
      ...Checkout
    }
  }
}
    fragment Checkout on Checkout {
  id
  totalPrice {
    gross {
      currency
      amount
    }
    tax {
      currency
      amount
    }
  }
  subtotalPrice {
    gross {
      currency
      amount
    }
  }
  quantity
  lines {
    id
    quantity
    variant {
      ...Variant
      product {
        ...ProductDetails
      }
    }
  }
}
fragment ProductDetails on Product {
  id
  slug
  name
  isAvailableForPurchase
  description
  seoTitle
  seoDescription
  pricing {
    priceRange {
      start {
        gross {
          currency
          amount
        }
      }
      stop {
        gross {
          currency
          amount
        }
      }
    }
  }
  media {
    url(size: 1080)
    type
    alt
  }
  collections {
    name
  }
  updatedAt
  variants {
    ...Variant
  }
}
fragment Variant on ProductVariant {
  id
  name
  attributes {
    attribute {
      slug
      name
      choices(first: 100) {
        edges {
          node {
            name
          }
        }
      }
    }
    values {
      name
    }
  }
  pricing {
    price {
      gross {
        currency
        amount
      }
    }
  }
}`) as unknown as TypedDocumentString<
  CheckoutDeleteLineMutation,
  CheckoutDeleteLineMutationVariables
>;